Inventory_management/requirements.txt
asgiref==3.9.1
axios==0.4.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
idna==3.10
lxml==6.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mysqlclient==2.2.7
Pygments==2.19.2
requests==2.32.4
rich==14.1.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
dj-database-url

